requirements.txt
python-telegram-bot==12.0.0b1
future>=0.16.0
certifi
tornado>=5.1
psycopg2>=2.8
